Merge pull request #11 from marcdalangin/dev Add gesture-based presentation script and presentation pages
This commit adds a new script, gesture_presentation.py, which allows users to control a presentation using hand gestures. The script uses a trained model to predict the hand gesture from a live video feed. The predicted gesture is then used to control the presentation, such as going to the next slide, the previous slide, or enabling draw mode.

In addition to the script, this commit also adds a set of presentation pages in image format to the presentation folder. These pages can be used to create a presentation that can be controlled using hand gestures.

To use the gesture-based presentation, users can simply run the gesture_presentation.py script and then start their presentation. The script will automatically detect the hand and its landmarks in the video feed and make a prediction of the hand gesture. The predicted gesture will then be used to control the presentation.